requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.21.0
pillow>=9.0.0
plotly>=5.10.0
fastai>=2.7.0
torch>=1.12.0
torchvision>=0.13.0
scikit-learn>=1.1.0
seaborn>=0.11.0
matplotlib>=3.5.0